Clarify subcircuit definition handling
Add DUT instance of spice_module to testbench.

Subcircuit definitions are pruned during DUT subcircuit parsing.

Pruned subcircuit definitions are written to a temporary file
with a testbench method export_subckts.

Attribute postlayout is set to True only id DSPF files are provided.
Otherwise it should be manually set. (Alternatively dedicated extraction
methods must be developed for different tools, but this is NOT a matter
of subcircuit definition parsing.)